refactoring: updated handlers structure.
