+++ telegram_bot/handlers/admin/__init__.py
@@ -1,0 +1,12 @@
+from . import block
+from . import message
+from . import users
+from . import user_detail
+
+
+__all__ = [
+    'block',
+    'message',
+    'users',
+    'user_detail',
+]
